camera: merge utilities into system
